Add way to get/set the default executor.
The default executor is the one that is used by default in the `graphql` and `execute` functions. This commit provides a way to override that executor to one configured to your liking. Additionally, since we are now doing that, the schema must be now passed to the `Executor.execute()` as the first argument, rather than as the first argument to `Executor()`.