Adding exception loggging
The error reporting code swallows exceptions in order to report them to
the client. When debugging (and when getting started with graphql as I
am) it is very useful to be able to see the exception stack traces.

I have therefore added a single logging statement to
`Executor.run_resolve_fn()`. Is it possible that logging is also
required elsewhere, I'm happy to take any advice this.