Add traceback to errors from failed promises (#240)
The stack attribute of GraphQLLocatedErrors was not set
for errors that resulted from asynchronous resolvers.

We add a unit test for this problem and a solution
that works with Python >= 3. In Python 2 this problem
is not yet solved, and may be never solved, because
dealing with stack traces is much more difficult there.