Revert the removal of the `_value` postfix (#234)
This restores the `_value(s)` prefix of three parameters of the
execute() and graphql() functions, which had been removed in v2.1,
in order to be compatible with graphql-core-next (v3), graphql-js,
and the older versions again.

The prefix-less parameters are still accepted, but now produce
deprecation warnings (in v2.1 it was the other way around).